+++ ChangeLog.md
@@ -5,6 +5,7 @@ QIIME 1.6.0-dev (changes since QIIME 1.6.0 go here)
 * Modified add_qiime_labels script to use normal metadata mapping file with a column specified for fasta file names to make more consistent with other scripts.
 * otu_category_significance.py now makes better use of the BIOM Table API, addressing a performance issue when using CSMat as the sparse backend.
 * Required biom-format version is now 1.1.2.
+* Added qiime.group.get_adjacent_distances, which is useful for plotting distances between "adjacent" sample ids in a list provided by the user. This is useful, for example, in plotting distances between adjacent temporal samples in a time series.
 
 QIIME 1.6.0 (18 Dec 2012)
 =========================